% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovarianceHelpers.R
\name{computeC}
\alias{computeC}
\title{Estimate C of Spiked Covariance Matrix Estimator}
\usage{
computeC(ell, p_n_ratio)
}
\arguments{
\item{p_n_ratio}{A \code{numeric} indicating the asymptotic ratio of the
number of features, p, and the number of observations, n. This ratio is
assumed to be between 0 and 1.}

\item{scaled_eig_vals}{A \code{numeric} vector of scaled estimated
eigenvalues.}

\item{p}{A \code{numeric} integer indicating the number of features in the
data.}
}
\value{
A \code{numeric} vector.
}
\description{
\code{computeC()} computes the c(ell) value described in
\insertCite{donoho2018;textual}{cvCovEst}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
