% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovarianceHelpers.R
\name{computeEll}
\alias{computeEll}
\title{Estimate Ell of Spiked Covariance Matrix Estimator}
\usage{
computeEll(scaled_eig_vals, p, p_n_ratio)
}
\arguments{
\item{scaled_eig_vals}{A \code{numeric} vector of scaled estimated
eigenvalues.}

\item{p}{A \code{numeric} integer indicating the number of features in the
data.}

\item{p_n_ratio}{A \code{numeric} indicating the asymptotic ratio of the
number of features, p, and the number of observations, n. This ratio is
assumed to be between 0 and 1.}
}
\value{
A \code{numeric} vector.
}
\description{
\code{computeEll()} computes the ell value described in
\insertCite{donoho2018;textual}{cvCovEst}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
