% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conditionNum}
\alias{conditionNum}
\title{Calculate Condition Number}
\usage{
conditionNum(estimate)
}
\arguments{
\item{estimate}{A symmetric \code{matrix} corresponding to a covariance
estimator.}
}
\value{
A \code{numeric} value indicating the condition number.  If undefined,
the condition number is set to 0.
}
\description{
\code{conditionNum()} calculates the condition number of a
covariance matrix estimator.
}
\keyword{internal}
