% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCovEst.R
\name{cvCovEst}
\alias{cvCovEst}
\title{Cross-Validated Covariance Matrix Estimator Selector}
\usage{
cvCovEst(
  dat,
  estimators = list(linearShrinkEst, thresholdingEst),
  estimator_params = list(linearShrinkEst = c(alpha = 0), thresholdingEst = c(gamma =
    0)),
  cv_scheme = "mc",
  mc_split = 0.5,
  v_folds = 10,
  boot_iter = 100,
  center = TRUE,
  scale = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{dat}{A numeric \code{data.frame} or matrix.}

\item{estimators}{A \code{list} of estimator functions.}

\item{estimator_params}{A named \code{list} of estimator parameters.}

\item{cv_scheme}{A \code{character} indicating the cross-validation scheme
to be emplyed. Defaults to Monte Carlo (\code{mc}).}

\item{mc_split}{A \code{numeric} between 0 and 1 indicating the proportion of
data in the validation set for each Monte Carlo iteration.}

\item{v_folds}{A \code{integer} larger than or equal to 1 indicating the
number of folds to use during cross-validation. The default is 10,
regardless of cross-validation scheme.}

\item{boot_iter}{A \code{integer} dictating the number of bootstrap
iterations used to compute the covariance terms of the cross-validated
scaled Frobenius loss. The default is set to 100.}

\item{center}{A \code{logical} indicating whether or not to center the
columns \code{dat}.}

\item{scale}{A \code{logical} indicating whether or not to scale the
columns \code{dat} to have variance 1.}

\item{parallel}{A \code{logical} option for whether to run the main
cross-validation loop with \code{\link[future.apply]{future_lapply}}.}
}
\value{
A \code{list} of results containing the following elements:
\itemize{
\item \code{estimate} - A \code{matrix} corresponding to the estimate of
the optimal covariance matrix estimator.
\item \code{estimator} - A \code{character} indicating the optimal
estimator and corresponding hyperparameters, if any.
\item \code{results_df} - A \code{data.frame} providing the results of
the cross-validation procedure.
}
}
\description{
\code{cvCovEst}
}
