% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvEigenPlot}
\alias{cvEigenPlot}
\title{Eigenvalue Plot}
\usage{
cvEigenPlot(dat, estimator, stat = "min", dat_orig)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{stat}{A string specifying a single summary statistic to use when
comparing two or more estimators.  Default is \code{'min'} for minimum
empirical risk.  Several stats can be used when only plotting one class of
estimator.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}
}
\value{
A plot of the leading eigenvalues for each estimator specified.
}
\description{
\code{cvEigenPlot} plots the eigenvalues of one or more
estimators produced by \code{cvCovEst}.
}
\keyword{external}
