% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{cvMatrixMetrics}
\alias{cvMatrixMetrics}
\title{Matrix Metrics for cvCovEst Object}
\usage{
cvMatrixMetrics(dat, dat_orig)
}
\arguments{
\item{dat}{A named list of class \code{"cvCovEst"} containing the
cross-validated risk assessment.}

\item{dat_orig}{The \code{numeric data.frame}, \code{matrix}, or similar
object originally passed to \code{\link{cvCovEst}()}.}
}
\value{
A grouped and ordered \code{\link[tibble]{tibble}} containing each
estimator candidate, the cross-validated risk, and the corresponding
metrics.  The output is grouped by estimator and ordered by the primary
hyperparameter if applicable.

@importFrom rlang exec .data
@importFrom dplyr bind_rows bind_cols group_by arrange \%>\%
}
\description{
\code{cvMatrixMetrics} computes various metrics and properties
for each covariance matrix estimator candidate passed to \code{cvCovEst}.
}
\keyword{internal}
