% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvRiskPlot}
\alias{cvRiskPlot}
\title{Empirical Risk Plot}
\usage{
cvRiskPlot(dat, est, conf = FALSE, plot_type = "risk")
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{conf}{A \code{logical} indicating whether confidence intervals should
be displayed for line plots.  Default is \code{FALSE}.}

\item{plot_type}{Argument passed to \code{theme_cvCovEst}.  Defaults to
\code{'risk'}}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}
}
\value{
A single plot or grid of plots for each estimator specified.
}
\description{
\code{cvRiskPlot} plots the empirical risk for a given estimator,
or set of estimators, as a function of the hyperparameters.
}
\keyword{internal}
