% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvSingleMelt}
\alias{cvSingleMelt}
\title{Single Heat Map Plot}
\usage{
cvSingleMelt(dat, estimator, stat, dat_orig)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{estimator}{A single string specifying which class of estimator to
visualize.}

\item{stat}{A single specifying the performance of the chosen estimator.  The
two options are \code{"best"} for the best performing estimator and
\code{"worst"} for the worst performing estimator.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}
}
\value{
A heat map plot of the desired covariance matrix estimator.
}
\description{
The \code{cvSingleMelt} plots a heat map visualization of a
single estimator for the covariance matrix.
}
\keyword{external}
