% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvSummary}
\alias{cvSummary}
\title{Summary Function for cvCovEst}
\usage{
cvSummary(dat, summary = "all")
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{summary}{A character vector specifying which summaries to output.  The
default is \code{'all'}.}
}
\value{
A named \code{list} where each element corresponds to the output of
of the requested summaries.
}
\description{
The \code{cvSummary} provides summary statistics regarding the
performance of \code{cvCovEst} and can be used for diagnostic plotting.
}
\keyword{external}
