% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{empRiskByClass}
\alias{empRiskByClass}
\title{Summary Statistics of Empirical Risk by Estimator Class}
\usage{
empRiskByClass(dat)
}
\arguments{
\item{dat}{The table of empirical risk calculations which is output by
\code{cvCovEst}.}
}
\value{
\code{\link[tibble]{tibble}} with rows corresponding to estimator
classes and columns corresponding to each summary statistic.
}
\description{
The \code{empRiskByClass} function calculates the following
summary statistics for the empirical risk within each class of estimator
passed to \code{cvCovEst}: minimum, Q1, median, mean, Q3, and maximum.  The
results are output as a \code{tibble}.
}
\keyword{internal}
