% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estAttributes}
\alias{estAttributes}
\title{Estimator Attributes Function}
\usage{
estAttributes(estimator)
}
\arguments{
\item{estimator}{A \code{character} vector specifying a class of estimator.}
}
\value{
A named \code{list} containing the attributes of the indicated
estimator.
}
\description{
\code{estAttributes()} returns a named list corresponding to the
attributes of a specific estimator implemented in the \code{cvCovEst}
package.
}
\keyword{internal}
