% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldFunctions.R
\name{frobeniusLoss}
\alias{frobeniusLoss}
\title{Frobenius Loss}
\usage{
frobeniusLoss(estimate, dat, rank_one_crossp)
}
\arguments{
\item{estimate}{A \code{matrix} corresponding to the estimate of the
covariance matrix.}

\item{dat}{A \code{data.frame} containing the data on which the loss will be
computed.}

\item{rank_one_crossp}{A \code{list} of \code{matrix} objects consisting
of the outer products of each of \code{dat}'s rows.}
}
\value{
The average Frobenius loss over \code{dat} of \code{estimate} as
a \code{numeric}.
}
\description{
\code{frobeniusLoss} computes the Frobenius
loss over the given dataset for the given covariance matrix estimate.
}
\keyword{internal}
