% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full-data-risk.R
\name{fullDataRisk}
\alias{fullDataRisk}
\title{Full Dataset Risk}
\usage{
fullDataRisk(dat, estimators, estimator_params, true_cov_mat)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{estimators}{A \code{list} of estimator functions to be
considered in the cross-validated selection procedure.}

\item{estimator_params}{A named \code{list} of arguments corresponding to
the hyperparameters of covariance matrix estimators in \code{estimators}.
The name of each list element should match the name of an estimator passed
to \code{estimators}. Each element of the \code{estimator_params} is itself
a named \code{list}, with the names corresponding to a given estimators'
hyperparameter(s). These hyperparameters may be in the form of a single
\code{numeric} or a \code{numeric} vector. If no hyperparameter is needed
for a given estimator, then the estimator need not be listed.}

\item{true_cov_mat}{A \code{matrix} like object representing the true
covariance matrix of the data generating distribution, which is assumed to
be Gaussian.}
}
\value{
A \code{\link[tibble]{tibble}} providing the full dataset risk of
each estimator.
}
\description{
\code{fullDataRisk} computes the full dataset risk under the
true data generating distribution, assuming that it is Gaussian.
}
\keyword{internal}
