% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getHypers}
\alias{getHypers}
\title{Hyperparameter Retrieval Function}
\usage{
getHypers(dat, summ_stat)
}
\arguments{
\item{dat}{A \code{data.frame} of estimators and their hyperparameter values.
Specifically, this is one of the outputs of \code{cvSummary}.}

\item{summ_stat}{A character vector specifying the summary statistic of
interest.}
}
\value{
A named \code{list} containing the names of all hyperparameters and
their associated values.
}
\description{
The \code{getHypers} retrieves the names and values of all
hyperparameters associated with an estimator passed to \code{cvCovEst}.
}
\keyword{internal}
