% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{hyperRisk}
\alias{hyperRisk}
\title{Summarize Cross-Validated Risks by Class with Hyperparameter}
\usage{
hyperRisk(dat)
}
\arguments{
\item{dat}{The \code{\link[tibble]{tibble}} of cross-validated risk calculations
which is output by \code{\link{cvCovEst}()}.}
}
\value{
A named \code{list} of data frames. Each list element corresponds to
a \code{\link[tibble]{tibble}} of summary statistics for a specific
estimator class. If no estimators have hyper-parameters, a message is
returned.
}
\description{
\code{hyperRisk()} groups together estimators of the
same class and parses the hyperparameter values over quantiles of the risk.
}
\keyword{internal}
