% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{multiHyperRisk}
\alias{multiHyperRisk}
\title{Multi-Hyperparameter Risk Plots}
\usage{
multiHyperRisk(
  dat,
  estimator,
  plot_type = "risk",
  switch_vars = FALSE,
  min_max = FALSE
)
}
\arguments{
\item{dat}{A data table of empirical risks.  Specifically, this is the
\code{risk_df} table output by \code{cvCovEst}.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{plot_type}{Argument passed to \code{theme_cvCovEst}.  Defaults to
\code{'risk'}.}

\item{switch_vars}{A \code{logical} indicating if the x-axis and factor
variables should be switched.  Default is \code{FALSE}.}

\item{min_max}{Default is \code{FALSE}.  If \code{TRUE}, only the minimum and
maximum values of the factor hyperparameter will be used.}
}
\value{
A plot of the empirical risk for the specified estimator
}
\description{
\code{multiHyperRisk} produces plots of the empirical risk for
estimators with more than one hyperparameter.  The function transforms one
of the hyperparameters into a factor and uses it to distinguish between the
risk of various estimators.
}
\keyword{internal}
