% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{nonlinearShrinkLWEst}
\alias{nonlinearShrinkLWEst}
\title{Non-Linear Shrinkage Estimator (Description Not Updated)}
\usage{
nonlinearShrinkLWEst(dat)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{nonlinearShrinkLWEst} invokes the analytical formula
presented by Ledoit and Wolfe for applying a nonlinear shrinkage function
to the sample eigenvalues of the covariance matrix.  The shrinkage function
utilizes the limiting spectral density of the sample eigenvalues and its
Hilbert Transform, both of which are representated using nonparametric
kernel estimation.  The result of the Hilbert transform is that smaller
clusters of eigenvalues are pushed up towards larger clusters, thus creating
a localized shrinkage effect rather than a global one.  For more
information, see \insertRef{Ledoit2018}{cvCovEst}.
}
