% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledFrobeniusLoss.R
\name{penScaledFrobeniusLoss}
\alias{penScaledFrobeniusLoss}
\title{Penalized Scaled Frobenius Loss}
\usage{
penScaledFrobeniusLoss(est_mat, sample_cov_mat, cov_sum)
}
\arguments{
\item{est_mat}{A \code{matrix} output from an estimator function.}

\item{sample_cov_mat}{A \code{matrix}, the sample covariance matrix.}

\item{cov_sum}{A \code{numeric} representing the sum of covariance estimates
between the identically positioned elements of \code{est_mat} and
\code{sample_cov_mat}.}
}
\value{
The \code{numeric} penalized scaled Frobenius loss of \code{est_mat}.
}
\description{
\code{penScaledFrobeniusLoss} computes the the penalized scaled
Frobenius matrix loss of a covariance matrix estimate.
}
\keyword{internal}
