% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_plots.R
\name{plotPoetEst}
\alias{plotPoetEst}
\title{Plot poetEst}
\usage{
plotPoetEst(dat, switch_vars = FALSE, min_max = FALSE)
}
\arguments{
\item{dat}{A data table of empirical risks.  Specifically, this is the
\code{risk_df} table output by \code{cvCovEst}.}

\item{switch_vars}{A \code{logical} indicating if the x-axis and factor
variables should be switched.  Default is \code{FALSE}.}

\item{min_max}{Default is \code{FALSE}.  If \code{TRUE}, only the minimum and
maximum values of the factor hyperparameter will be used.}
}
\value{
A plot object
}
\description{
\code{plotPoetEst} performs actions specific to plotting
the empirical risk of the POET estimator.
}
\keyword{internal}
