% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_plots.R
\name{plotRobustPoetEst}
\alias{plotRobustPoetEst}
\title{Plot robustPoetEst}
\usage{
plotRobustPoetEst(dat, switch_vars = FALSE, min_max = FALSE)
}
\arguments{
\item{dat}{A data table of empirical risks.  Specifically, this is the
\code{risk_df} table output by \code{\link{cvCovEst}()}.}

\item{switch_vars}{A \code{logical} indicating if the x-axis and factor
variables should be switched.  Default is \code{FALSE}.}

\item{min_max}{A \code{logical}. Default is \code{FALSE}. If \code{TRUE},
only the minimum and maximum values of the factor hyperparameter will be
used.}
}
\value{
A list of plots
}
\description{
\code{plotRobustPoetEst()} performs actions specific to plotting
the empirical risk of the Robust POET estimator.
}
\keyword{internal}
