% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{robustPoetEst}
\alias{robustPoetEst}
\title{Robust POET Estimator for Elliptical Distributions}
\usage{
robustPoetEst(Y, k, lambda, var_estimation)
}
\arguments{
\item{Y}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{k}{An \code{integer} indicating the number of unobserved latent
factors. Empirical evidence suggests that the POET estimator is robust to
overestimations of this hyperparameter \insertCite{fan2013}{cvCovEst}. In
practice, it is therefore preferable to use larger values.}

\item{lambda}{A non-negative \code{numeric} defining the amount of
thresholding applied to each element of sample covariance matrix's
orthogonal complement.}

\item{var_estimation}{A \code{character} string giving the method of
variance estimation method. This must be one of the strings "\code{mad}",
"\code{sample}", "\code{huber}".}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{robustPoetEst} implements the robust version of Principal
Orthogonal complEment Thresholding (POET) estimator, a nonparametric,
unobserved-factor-based estimator of the covariance matrix when the
underlying distribution is ellipitcal\insertCite{fan2018}{cvCovEst}. The
estimator is defined as the sum of the sample covariance matrix'
rank-\code{k} approximation and its thresholded principal orthogonal
complement. Rank-\code{k} approximation is constructed from the sample
covariance matrix, its leading eigenvalues, and leading eigenvectors,
first two estimated by M-estimation and marginal Kendall's tau estimator
while the last one estimated by spatial Kendall's tau estimator. The hard
thresholding function is used here, though others could be used instead.
}
\references{
\insertAllCited{}
}
