% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{sampleCovEst}
\alias{sampleCovEst}
\title{Sample Covariance Matrix}
\usage{
sampleCovEst(dat)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{sampleCovEst} computes the sample covariance matrix. This
function is a simple wrapper around \code{\link[coop]{covar}}.
}
\examples{
sampleCovEst(dat = mtcars)

}
