% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledFrobeniusLoss.R
\name{scaledFrobeniusLoss}
\alias{scaledFrobeniusLoss}
\title{Scaled Frobenius Loss}
\usage{
scaledFrobeniusLoss(est_mat, sample_cov_mat)
}
\arguments{
\item{est_mat}{A \code{matrix} output from an estimator function.}

\item{sample_cov_mat}{A \code{matrix}, the sample covariance matrix.}
}
\value{
The \code{numeric} scaled Frobenius loss of \code{est_mat}.
}
\description{
\code{scaledFrobeniusLoss} computes the scaled Frobenius matrix
loss of a covariance matrix estimate.
}
\keyword{internal}
