% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCovEst.R
\name{strToNumber}
\alias{strToNumber}
\title{Convert String to Numeric or Integer When Needed}
\usage{
strToNumber(x)
}
\arguments{
\item{x}{A \code{character} representing a number or an integer.}
}
\value{
\code{x} converted to the appropriate type.
}
\description{
Convert String to Numeric or Integer When Needed
}
\keyword{internal}
