% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunctions.R
\name{sumNaiveBootstrap}
\alias{sumNaiveBootstrap}
\title{Naive Bootstrap Estimator of Second Term in Scaled Frobenius Loss}
\usage{
sumNaiveBootstrap(
  estimator_fun,
  estimates,
  sample_cov_mat,
  train_data,
  valid_data,
  num_iter,
  est_args = NULL
)
}
\arguments{
\item{estimator_fun}{The covariance matrix estimator to be applied to the
resampled training data.}

\item{estimates}{A \code{matrix} of the estimator's covariance matrix
estimate over a given fold.}

\item{sample_cov_mat}{A \code{matrix} of the sample covariance matrix over a
given fold.}

\item{train_data}{A \code{data.frame} of the training data.}

\item{valid_data}{A \code{data.frame} of the validation data.}

\item{num_iter}{A positive integer defining the number of bootstrap samples
to compute.}

\item{est_args}{Arguments corresponding to the hyperparameters of the
covariance matrix estimator, \code{estimator_fun}.}
}
\value{
A \code{numeric} estimating the second term of the scaled Frobenius
loss.
}
\description{
\code{sumNaiveBootstrap} estimates the second term in scaled
Frobenius loss by computing the bootstrap means of the entries in the
covariance matrix  estimator and sample covariance matrix.
}
\keyword{internal}
