% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingFunctions.R
\name{sumNaiveCovBootstrap}
\alias{sumNaiveCovBootstrap}
\title{Naive Bootstrap Estimator of Covariance Matrix and Sample Covariance Matrix Covariance}
\usage{
sumNaiveCovBootstrap(
  estimator_fun,
  train_data,
  valid_data,
  num_iter,
  est_args = NULL
)
}
\arguments{
\item{estimator_fun}{The covariance matrix estimator to be applied to the
resampled training data.}

\item{train_data}{A \code{data.frame} of the training data.}

\item{valid_data}{A \code{data.frame} of the validation data.}

\item{num_iter}{A positive integer defining the number of bootstrap samples
to compute.}

\item{est_args}{Arguments corresponding to the hyperparameters of the
covariance matrix estimator, \code{estimator_fun}.}
}
\value{
A \code{numeric} representing the sum of all elements in the
estimated covariance matrix of the \code{estimator_fun} and the sample
covariance matrix.
}
\description{
\code{sumNaiveCovBootstrap} estimates the covariance matrix
between a generic covariance matrix estimator and the sample covariance
matrix fit over a training set and a validation set, respectively. The
sum of the estimated covariance matrix's elements are returned.
}
\keyword{internal}
