% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding-functions.R
\name{symmetricApply}
\alias{symmetricApply}
\title{Symmetric Apply Function for Covariance Matricies}
\usage{
symmetricApply(dat, sym_fun, sym_args)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{sym_fun}{a function to apply to each element of the covariance matrix.}

\item{sym_args}{a named vector or \code{list} of arguments to be passed to
\code{sym_fun}.}
}
\value{
A \code{matrix}.
}
\description{
\code{symmetricApply} implements a version of \code{apply}
expressly made for looping over symmetric matrices.
}
\keyword{internal}
