% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{taperingEst}
\alias{taperingEst}
\title{Tapering Estimator}
\usage{
taperingEst(dat, k)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{k}{a non-negative, even numeric integer}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matirx.
}
\description{
\code{taperingEst} estimates the covariance matrix of a data
frame with ordered variables by gradually shrinking each off-diagonal entry
of the sample covariance matrix towards zero depending on its distance
from the diagonal.  The i, j - entry of the estimated covariance matrix will
be multipled by either 1, 0, or a given weight depending on a distance
parameter, k.  The value of k also determines the size of the weight.
}
