% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_cvCovEst}
\alias{theme_cvCovEst}
\title{cvCovEst Plot Theme}
\usage{
theme_cvCovEst(plot_type, ...)
}
\arguments{
\item{plot_type}{A \code{character} vector specifying which plot is to be
displayed. Can contain more than one value in the case of
\code{plot_type = c("heatmap", "summary")}.}
}
\value{
A \code{ggplot} theme.
}
\description{
\code{theme_cvCovEst()} defines the overall theme of the
\code{cvCovEst} package plotting functions and makes changes depending on
which plot function is being called.
}
\keyword{internal}
