% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldFunctions.R
\name{trueFrobeniusLoss}
\alias{trueFrobeniusLoss}
\title{True Cross-Validated Frobenius Loss}
\usage{
trueFrobeniusLoss(estimate, true_covar)
}
\arguments{
\item{estimate}{A \code{matrix} corresponding to the estimate of the
covariance matrix.}

\item{true_covar}{A \code{matrix} corresponding to the true covariance matrix
of the data generating distribution.}
}
\value{
The true average Frobenius loss over the validation dataset of
\code{estimate} as a \code{numeric}.
}
\description{
\code{trueFrobeniusLoss} computes the true cross-validated
Frobenius loss over the validation dataset.
}
\keyword{internal}
