% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{bestInClass}
\alias{bestInClass}
\title{Showing Best Estimator Within Each Class of Estimators}
\usage{
bestInClass(dat)
}
\arguments{
\item{dat}{The table of empirical risk calculations which is output by
\code{cvCovEst}.}
}
\value{
A \code{data.frame} with rows corresponding to estimator classes and
columns for hyperparameter values and empirical risk for the best estimator
in that class.
}
\description{
The \code{bestInClass} function finds the best performing
estimator within each class of estimator passed to \code{cvCovEst} and
finds the associated hyperparameters if applicable.
}
\keyword{internal}
