% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{bestInClass}
\alias{bestInClass}
\title{Showing Best Estimator Within Each Class of Estimators}
\usage{
bestInClass(dat, worst = FALSE)
}
\arguments{
\item{dat}{The \code{\link[tibble]{tibble}} of cross-validated risks which is
output by \code{\link{cvCovEst}()}.}

\item{worst}{This facilitates the option to choose the worst performing
estimator in each class.  Default is \code{FALSE}.}
}
\value{
\code{\link[tibble]{tibble}} with rows corresponding to estimator
classes and columns for hyperparameter values, cross-validated risk, and
other summary metrics for the best (or worst) estimator in that class.
}
\description{
\code{bestInClass()} finds the best performing estimator within
each class of estimator passed to \code{\link{cvCovEst}()} and
finds the associated hyper-parameters if applicable.
}
\keyword{internal}
