% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArgs.R
\name{checkArgs}
\alias{checkArgs}
\title{Check Arguments Passed to cvCovEst}
\usage{
checkArgs(
  dat,
  estimators,
  estimator_params,
  cv_loss,
  cv_scheme,
  mc_split,
  v_folds,
  center,
  scale,
  parallel,
  true_cov_mat = NULL
)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{estimators}{A \code{list} of estimator functions to be
considered in the cross-validated selection procedure.}

\item{estimator_params}{A named \code{list} of arguments corresponding to
the hyperparameters of covariance matrix estimators in \code{estimators}.
The name of each list element should match the name of an estimator passed
to \code{estimators}. Each element of the \code{estimator_params} is itself
a named \code{list}, with the names corresponding to a given estimators'
hyperparameter(s). These hyperparameters may be in the form of a single
\code{numeric} or a \code{numeric} vector. If no hyperparameter is needed
for a given estimator, then the estimator need not be listed.}

\item{cv_loss}{A \code{function} indicating the loss function to use.
Defaults to the scaled Frobenius loss, \code{cvFrobeniusLoss}.
The matrix-based version, \code{cvMatrixFrobeniusLoss} is offered as well.}

\item{cv_scheme}{A \code{character} indicating the cross-validation scheme
to be employed. There are two options: (1) V-fold cross-validation, via
\code{"v_folds"}; and (2) Monte Carlo cross-validation, via \code{"mc"}.
Defaults to Monte Carlo cross-validation.}

\item{mc_split}{A \code{numeric} between 0 and 1 indicating the proportion
of data in the validation set of each Monte Carlo cross-validation fold.}

\item{v_folds}{A \code{integer} larger than or equal to 1 indicating the
number of folds to use during cross-validation. The default is 10,
regardless of cross-validation scheme.}

\item{center}{A \code{logical} indicating whether or not to center the
columns of \code{dat}.}

\item{scale}{A \code{logical} indicating whether or not to scale the
columns of \code{dat} to have variance 1.}

\item{parallel}{A \code{logical} option indicating whether to run the main
cross-validation loop with \code{\link[future.apply]{future_lapply}}. This
is passed directly to \code{\link[origami]{cross_validate}}.}

\item{true_cov_mat}{A \code{matrix} like object representing the true
covariance matrix of the data generating distribution, which is assumed to
be Gaussian. This parameter is intended for use only in simulation studies,
and defaults to a value of \code{NULL}. If not \code{NULL}, the
cross-validated conditional risk difference ratio of the estimator selected
by \code{cvCovEst} is computed relative to the cross-validated oracle
selector.}
}
\value{
Whether all argument conditions are satisfied
}
\description{
The \code{checkArgs} function verifies that all arguments passed
to the \code{cvCovEst} function meet its specifications.
}
\keyword{internal}
