% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{checkPlotSumArgs}
\alias{checkPlotSumArgs}
\title{Check Arguments Passed to plot.cvCovEst and summary.cvCovEst}
\usage{
checkPlotSumArgs(
  dat,
  dat_orig,
  which_fun,
  estimator,
  plot_type,
  summ_fun,
  stat,
  conf,
  k,
  leading
)
}
\arguments{
\item{dat}{An object of class, "cvCovEst".  Specifically, this is the
standard output of the function \code{cvCovEst}.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}

\item{which_fun}{A choice of \code{'plot'} or \code{'summary'} depending on
which function is being checked.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{plot_type}{A character vector specifying one of four choices of
diagnostic plots.  Default is \code{'summary'}.  See Details for more about
each plotting choice.}

\item{summ_fun}{A character vector specifying which summaries to output.}

\item{stat}{A character vector of one or more summary statistics to use when
comparing estimators.  Default is \code{'min'} for minimum
empirical risk.  See Details for more options.}

\item{conf}{A \code{logical} indicating whether or not to include confidence
bands in empirical risk plots.  Default is \code{FALSE}.}

\item{k}{The number of leading/trailing eigenvalues to plot. If \code{NULL},
will default to the number of columns in \code{dat_orig}.}

\item{leading}{A \code{logical} indicating if the leading eigenvalues should
be used.  Default is \code{TRUE}.  If \code{FALSE}, the trailing eigenvalues
are used instead.}
}
\value{
Whether all argument conditions are satisfied
}
\description{
The \code{checkPlotSumArgs} function verifies that all arguments
passed to the \code{plot.cvCovEst} and \code{summary.cvCovEst} functions meet
their specifications.  Some additional arguments may be checked at the
individual function level.
}
\keyword{internal}
