% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCovEst.R
\name{cvCovEst}
\alias{cvCovEst}
\title{Cross-Validated Covariance Matrix Estimator Selector}
\usage{
cvCovEst(
  dat,
  estimators = c(linearShrinkEst, thresholdingEst, sampleCovEst),
  estimator_params = list(linearShrinkEst = list(alpha = 0), thresholdingEst =
    list(gamma = 0)),
  cv_loss = cvFrobeniusLoss,
  cv_scheme = "v_fold",
  mc_split = 0.5,
  v_folds = 10L,
  center = TRUE,
  scale = FALSE,
  parallel = FALSE,
  true_cov_mat = NULL
)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{estimators}{A \code{list} of estimator functions to be
considered in the cross-validated selection procedure.}

\item{estimator_params}{A named \code{list} of arguments corresponding to
the hyperparameters of covariance matrix estimators in \code{estimators}.
The name of each list element should match the name of an estimator passed
to \code{estimators}. Each element of the \code{estimator_params} is itself
a named \code{list}, with the names corresponding to a given estimators'
hyperparameter(s). These hyperparameters may be in the form of a single
\code{numeric} or a \code{numeric} vector. If no hyperparameter is needed
for a given estimator, then the estimator need not be listed.}

\item{cv_loss}{A \code{function} indicating the loss function to use.
Defaults to the scaled Frobenius loss, \code{cvFrobeniusLoss}.
The matrix-based version, \code{cvMatroxFrobeniusLoss} is offered as well.}

\item{cv_scheme}{A \code{character} indicating the cross-validation scheme
to be employed. There are two options: (1) V-fold cross-validation, via
\code{"v_folds"}; and (2) Monte Carlo cross-validation, via \code{"mc"}.
Defaults to V-fold cross-validation.}

\item{mc_split}{A \code{numeric} between 0 and 1 indicating the proportion
of data in the validation set of each Monte Carlo cross-validation fold.}

\item{v_folds}{A \code{integer} larger than or equal to 1 indicating the
number of folds to use during cross-validation. The default is 10,
regardless of cross-validation scheme.}

\item{center}{A \code{logical} indicating whether or not to center the
columns of \code{dat}. Set to \code{FALSE} only if the columns have already
been centered. Defaults to \code{TRUE}.}

\item{scale}{A \code{logical} indicating whether or not to scale the
columns of \code{dat} to have variance 1. Defaults to \code{FALSE}.}

\item{parallel}{A \code{logical} option indicating whether to run the main
cross-validation loop with \code{\link[future.apply]{future_lapply}}. This
is passed directly to \code{\link[origami]{cross_validate}}.}

\item{true_cov_mat}{A \code{matrix} like object representing the true
covariance matrix of the data generating distribution, which is assumed to
be Gaussian. This parameter is intended for use only in simulation studies,
and defaults to a value of \code{NULL}. If not \code{NULL}, various
conditional risk difference ratios of the estimator selected
by \code{cvCovEst} are computed relative to the different oracle selectors.}
}
\value{
A \code{list} of results containing the following elements:
\itemize{
\item \code{estimate} - A \code{matrix} corresponding to the estimate of
the optimal covariance matrix estimator.
\item \code{estimator} - A \code{character} indicating the optimal
estimator and corresponding hyperparameters, if any.
\item \code{risk_df} - A \code{\link[tibble]{tibble}} providing the
cross-validated risk estimates of each estimator.
\item \code{cv_df} - A \code{\link[tibble]{tibble}} providing
each estimators' loss over the various folds of the cross-validatied
procedure.
\item \code{args} - A named \code{list} containing arguments passed to
\code{cvCovEst}.

}
If the true covariance matrix is input through \code{true_cov_mat}, then
three additional elements are returned:
\itemize{
\item \code{cv_cv_riskdiff} - A \code{numeric} corresponding
to the conditional cross-validated risk difference of the
cvCovEst selection.
\item \code{oracle_cv_riskdiff} - A \code{numeric}
corresponding to the conditional cross-validated risk difference of the
oracle selection.
\item \code{cv_oracle_riskdiff_ratio} - A \code{numeric} corresponding
to the cross-validated risk difference ratio of the
cvCovEst selection and the cross-validated dataset oracle selection.
\item \code{full_data_riskdiff_ratio} - A \code{numeric} corresponding
to the full data risk difference ratio of the
cvCovEst selection and the full dataset oracle selection.
}
}
\description{
\code{cvCovEst} identifies the optimal covariance matrix
estimator from among a set of candidate estimators.
}
