% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvEigenPlot}
\alias{cvEigenPlot}
\title{Eigenvalue Plot}
\usage{
cvEigenPlot(
  dat,
  estimator,
  stat = "min",
  dat_orig,
  k,
  leading = TRUE,
  plot_type = "eigen",
  cv_details,
  has_hypers
)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{stat}{A string specifying a single summary statistic to use when
comparing two or more estimators.  Default is \code{'min'} for minimum
empirical risk.  Several stats can be used when only plotting one class of
estimator.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}

\item{k}{The number of eigenvalues to plot.  Must be less than or equal to
the number of columns of the original data matrix.}

\item{leading}{A \code{logical} indicating if the leading eigenvalues should
be used.  Defautlt is \code{TRUE}.  If \code{FALSE}, the trailing eigenvalues
will be used instead.}

\item{plot_type}{Argument passed to \code{theme_cvCovEst}.  Defaults to
\code{'eigen'}.}

\item{cv_details}{Character vector summarizing key arguments passed to
\code{cvCovEst}.}

\item{has_hypers}{A character vector containing the names of current
estimators with hyperparameters.}
}
\value{
A plot, or grid of plots, showing the k leading or trailing
eigenvalues of the specified estimators and associated summary statistics of
the empirical risk.
}
\description{
\code{cvEigenPlot} plots the eigenvalues of one or more
estimators produced by \code{cvCovEst}.
}
\keyword{internal}
