% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldFunctions.R
\name{cvFrobeniusLoss}
\alias{cvFrobeniusLoss}
\title{Cross-Validation Function for Scaled Frobenius Loss}
\usage{
cvFrobeniusLoss(fold, dat, resample_cov_fun, resample_iter, estimator_fun, ...)
}
\arguments{
\item{fold}{A \code{fold} object over which the estimation procedure will
take place.}

\item{dat}{The full \code{data.frame} on which the cross-validated procedure
is performed.}

\item{resample_iter}{A \code{numeric} indicating the number of repetitions
to be performed by \code{resample_fun}.}

\item{estimator_fun}{The covariance matrix estimator to be applied to the
training data.}

\item{...}{Arguments corresponding to the hyperparameters of the
covariance matrix estimator, \code{estimator_fun}.}

\item{resample_fun}{The function defining the resampling-based procedure used
to estimate the covariances of entries in the covariance matrix estimator
and the sample covariance matrix.}
}
\value{
A \code{list} providing information on the estimator, its
hyperparameters (if any), and scaled Frobenius loss for the
given \code{fold}.
}
\description{
\code{cvFrobeniusLoss} evaluates the scaled Frobenius loss over
a \code{fold} object defined via the \pkg{origami} package
\insertRef{Coyle2018}{cvCovEst}.
}
\keyword{internal}
