% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{cvMatrixMetrics}
\alias{cvMatrixMetrics}
\title{Matrix Metrics for cvCovEst Object}
\usage{
cvMatrixMetrics(object, dat_orig)
}
\arguments{
\item{object}{A named list of class \code{"cvCovEst"} containing the
cross-validated risk assessment.}

\item{dat_orig}{The \code{numeric data.frame}, \code{matrix}, or similar
object originally passed to \code{\link{cvCovEst}()}.}
}
\value{
A named list of class \code{"cvCovEst"} whose cross-validated risk
assessment is now a \code{\link[tibble]{tibble}} containing the
corresponding metrics for each estimate.  The \code{\link[tibble]{tibble}}
is grouped by estimator and ordered by the primary hyperparameter if
applicable.
}
\description{
\code{cvMatrixMetrics} computes various metrics and properties
for each covariance matrix estimator candidate's estimate.
}
\keyword{internal}
