% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvMultiMelt}
\alias{cvMultiMelt}
\title{Multiple Heat Map Plot}
\usage{
cvMultiMelt(
  dat,
  estimator,
  single_stat = TRUE,
  stat = "best",
  multi_stat = "1",
  dat_orig
)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{single_stat}{A \code{logical} indicating whether to compare estimators
based on one summary statistic, such as the best performing estimator.  The
default is \code{TRUE}.  If set to \code{FALSE}, then multiple comparisons
are allowed.  See Details.}

\item{stat}{A string specifying the single summary statistic to use when
comparing two or more estimators.  See Details.}

\item{multi_stat}{A character indicating the type of comparison method to use.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}
}
\value{
A grid of heat map plots comparing the desired covariance matrix
estimators.
}
\description{
The \code{cvMultiMelt} compares the structure of two or more
covariance matrix estimators through a grid of heat maps, where each heat map
corresponds to a different estimator.
}
\keyword{external}
