% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvRiskPlot}
\alias{cvRiskPlot}
\title{Empirical Risk Plot}
\usage{
cvRiskPlot(
  dat,
  est,
  plot_type = "risk",
  cv_details,
  switch_vars = FALSE,
  min_max = FALSE
)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{plot_type}{Argument passed to \code{theme_cvCovEst}.  Defaults to
\code{"risk"}}

\item{cv_details}{Character vector summarizing key arguments passed to
\code{cvCovEst}.}

\item{switch_vars}{Default is \code{FALSE}. If \code{TRUE},
the hyperparameters used for the x-axis and factor variables are switched.
Only applies to estimators with more than one hyperparameter.}

\item{min_max}{Default is \code{FALSE}.  If \code{TRUE}, only the minimum and
maximum values of the factor hyperparameter will be used.  Only applies to
estimators with more than one hyperparameter.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}
}
\value{
A single plot or grid of plots for each estimator specified.
}
\description{
\code{cvRiskPlot} plots the empirical risk for a given
estimator, or set of estimators, as a function of the hyperparameters.
}
\keyword{internal}
