% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cvSummaryPlot}
\alias{cvSummaryPlot}
\title{Summary Plot}
\usage{
cvSummaryPlot(
  dat,
  estimator,
  dat_orig,
  stat,
  conf,
  k,
  leading,
  plot_type = "summary",
  cv_details,
  has_hypers
)
}
\arguments{
\item{dat}{A named \code{list}.  Specifically, this is the standard output of
\code{cvCovEst}.}

\item{estimator}{A character vector specifying which class of estimator to
plot.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}

\item{plot_type}{Argument passed to \code{theme_cvCovEst}.  Defaults to
\code{'summary'}.}

\item{cv_details}{Character vector summarizing key arguments passed to
\code{cvCovEst}.}

\item{...}{Additional arguments to be passed to \code{cvCovEst} plotting
functions.}
}
\value{
A collection of plots and summary statistics.
}
\description{
\code{cvSummaryPlot} combines plots of the empirical risk,
eigenvalues, and covariance heat map of a given estimator class together into
one plot along with a table of summary statistics.
}
\keyword{internal}
