% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledFrobeniusLoss.R
\name{frobeniusloss}
\alias{frobeniusloss}
\title{Aggregated Frobenius Loss}
\usage{
frobeniusloss(est_mat, valid_data)
}
\arguments{
\item{est_mat}{A \code{matrix} output from an estimator function.}

\item{valid_data}{A \code{numeric} \code{matrix} corresponding to the
validation set.}
}
\value{
The \code{numeric} aggregated Frobenius loss of \code{est_mat} for
over \code{valid_data}.
}
\description{
\code{frobeniusLoss} computes the aggregated, cross-validated
loss of \code{est_mat}, an estimate output by a covariance matrix
estimator, over a validation set.
}
\keyword{internal}
