% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hyperRisk}
\alias{hyperRisk}
\title{Summarize Risk by Class with Hyperparameter}
\usage{
hyperRisk(dat)
}
\arguments{
\item{dat}{The table of empirical risk calculations which is output by
\code{cvCovEst}.}
}
\value{
A named \code{list} of data frames.  Each list element corresponds to
a \code{data.frame} of summary statistics for a specific estimator class.
If no estimators have hyper-parameters, a message is returned.
}
\description{
The \code{hyperRisk} function groups together estimators of the
same class and parses the hyperparameter values over quantiles of the risk.
}
\keyword{internal}
