% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.cvCovEst}
\alias{is.cvCovEst}
\title{Check for cvCovEst Class}
\usage{
is.cvCovEst(x)
}
\arguments{
\item{x}{The specific object to test.}
}
\value{
A \code{logical} indicating \code{TRUE} if \code{x} inherits from
class \code{cvCovEst}.
}
\description{
\code{is.cvCovEst} provides a generic method for checking if
input is of class \code{cvCovEst}.
}
