% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{matrixMetrics}
\alias{matrixMetrics}
\title{Calculate Matrix Metrics}
\usage{
matrixMetrics(estimate)
}
\arguments{
\item{estimate}{A symmetric \code{matrix} corresponding to a covariance
estimator.}
}
\value{
A named \code{list} containing the different matrix metrics.  Note:
if the smallest eigenvalue of the matrix is 0, the condition number is set
to 0.
}
\description{
\code{matrixMetrics()} calculates the condition number, the
sparsity percentage, and the sign of a covariance matrix estimator.
}
\keyword{internal}
