% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{nonlinearShrinkLWEst}
\alias{nonlinearShrinkLWEst}
\title{Analytical Non-Linear Shrinkage Estimator}
\usage{
nonlinearShrinkLWEst(dat)
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{nlShrinkLWEst} invokes the analytical estimator
presented by \insertRef{Ledoit2020}{cvCovEst} for applying a nonlinear
shrinkage function to the sample eigenvalues of the covariance matrix.
The shrinkage function relies on an application of the Hilbert Transform to
an estimate of the sample eigenvalues' limiting spectral density. This
estimated density is computed with the Epanechnikov kernel using a global
bandwidth parameter of n^(-1/3). The resulting shrinkage function pulls
eigenvalues towards the nearest mode of their empirical distribution, thus
creating a localized shrinkage effect rather than a global one.
}
