% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.cvCovEst}
\alias{plot.cvCovEst}
\title{Generic Plot Method for cvCovest}
\usage{
\method{plot}{cvCovEst}(
  x,
  dat_orig,
  estimator,
  plot_type = c("summary"),
  stat = c("min"),
  k = NULL,
  leading = TRUE,
  abs_v = TRUE,
  switch_vars = FALSE,
  min_max = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class, "cvCovEst".  Specifically, this is the
standard output of the function \code{cvCovEst}.}

\item{dat_orig}{The numeric \code{data.frame}, \code{matrix}, or similar
object originally passed to \code{cvCovEst}.}

\item{estimator}{A character vector specifying one or more classes of
estimators to compare.}

\item{plot_type}{A character vector specifying one of four choices of
diagnostic plots.  Default is \code{'summary'}.  See Details for more about
each plotting choice.}

\item{stat}{A character vector of one or more summary statistics to use when
comparing estimators.  Default is \code{'min'} for minimum
empirical risk.  See Details for more options.}

\item{k}{The number of leading/trailing eigenvalues to plot. If \code{NULL},
will default to the number of columns in \code{dat_orig}.}

\item{leading}{A \code{logical} indicating if the leading eigenvalues should
be used.  Default is \code{TRUE}.  If \code{FALSE}, the trailing eigenvalues
are used instead.}

\item{abs_v}{A code{logical} determining if the absolute value of the matrix
entries should be used for plotting the matrix heatmap.  Default is \code{TRUE}.}

\item{switch_vars}{Default is \code{FALSE}. If \code{TRUE}, the hyperparameters
used for the x-axis and factor variables are switched.  Only applies to
estimators with more than one hyperparameter.}

\item{min_max}{Default is \code{FALSE}.  If \code{TRUE}, only the minimum and
maximum values of the factor hyperparameter will be used.  Only applies to
estimators with more than one hyperparameter.}

\item{...}{Additional arguments}
}
\value{
A plot object
}
\description{
\code{plot.cvCovEst} is a generic plotting method for objects of
class, "cvCovEst".  The method is designed as a tool for diagnostic and
exploratory analysis purposes when selecting a covariance matrix estimator.
}
