% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{robustPoetEst}
\alias{robustPoetEst}
\title{Robust POET Estimator for Elliptical Distributions}
\usage{
robustPoetEst(dat, lambda, k, var_estimation = c("mad", "sample", "huber"))
}
\arguments{
\item{dat}{A numeric \code{data.frame}, \code{matrix}, or similar object.}

\item{lambda}{A non-negative \code{numeric} defining the amount of
thresholding applied to each element of sample covariance matrix's
orthogonal complement.}

\item{k}{An \code{integer} indicating the number of unobserved latent
factors. Empirical evidence suggests that the POET estimator is robust to
overestimations of this hyperparameter \insertCite{fan2013}{cvCovEst}. In
practice, it is therefore preferable to use larger values.}

\item{var_estimation}{A \code{character} string dictating which
variance estimator to use. This must be one of the strings
"\code{mad}", "\code{sample}", or "\code{huber}".  "\code{mad}" estimates
variances via median absolute deviation; "\code{sample}" uses sample
variances; "\code{huber}" uses M-estimator for variance under Huber loss.
Defaults to "\code{sample}".}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{robustPoetEst} implements the robust version of Principal
Orthogonal complEment Thresholding (POET) estimator, a nonparametric,
unobserved-factor-based estimator of the covariance matrix when the
underlying distribution is ellipitcal \insertCite{fan2018}{cvCovEst}. The
estimator is defined as the sum of the sample covariance matrix'
rank-\code{k} approximation and its thresholded principal orthogonal
complement. The rank-\code{k} approximation is constructed from the sample
covariance matrix, its leading eigenvalues, and leading eigenvectors.
The sample covariance matrix and leading eigenvalues are initially
estimated via M-estimation and the marginal Kendall's tau estimator.
The leading eigenvectors are estimated using spatial Kendall's tau
estimator. The hard thresholding function is used to regularize the
idiosyncratic errors' estimated covariance matrix, though other
regularization schemes could beused instead.
}
\references{
\insertAllCited{}
}
