% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdingFunctions.R
\name{scadThreshold}
\alias{scadThreshold}
\title{Smoothly Clipped Absolute Deviation Thresholding Function}
\usage{
scadThreshold(entry, lambda, a)
}
\arguments{
\item{entry}{A \code{numeric} entry in a covariance matrix estimate.}

\item{lambda}{A non-negative \code{numeric} defining the degree of
thresholding applied to each element of \code{dat}'s sample covariance
matrix.}

\item{a}{A \code{numeric} larger than or equal to \code{2} defining the
point at which the SCAD thresholding functions becomes equal to the hard
thresholding function.}
}
\value{
A regularized \code{numeric}.
}
\description{
\code{scadThreshold()} applies the smoothly clipped absolute
deviation thresholding function to the entries of a matrix.
}
\keyword{internal}
