% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledFrobeniusLoss.R
\name{scaledFrobeniusLoss}
\alias{scaledFrobeniusLoss}
\title{Scaled Frobenius Loss}
\usage{
scaledFrobeniusLoss(est_mat, sample_cov_mat, cov_sum)
}
\arguments{
\item{est_mat}{A \code{matrix} output from an estimator function.}

\item{sample_cov_mat}{A \code{matrix} consisting of the sample covariance
matrix.}

\item{cov_sum}{A \code{numeric} representing the sum of covariance estimates
between the identically positioned elements of \code{est_mat} and
\code{sample_cov_mat}.}
}
\value{
The \code{numeric} scaled Frobenius loss of \code{est_mat}.
}
\description{
\code{scaledFrobeniusLoss} computes the tractable version of the
scaled Frobenius loss for an estimate of the covariance matrix.
}
\keyword{internal}
