% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary.cvCovEst}
\alias{summary.cvCovEst}
\title{Summary Function for cvCovEst}
\usage{
\method{summary}{cvCovEst}(
  dat,
  summ_fun = c("empRiskByClass", "bestInClass", "worstInClass", "hyperRisk"),
  ...
)
}
\arguments{
\item{dat}{A named \code{list} of class \code{"cvCovEst"}.}

\item{summ_fun}{A character vector specifying which summaries to output.}

\item{...}{Additional arguments passed to summary method.}
}
\value{
A named \code{list} where each element corresponds to the output of
of the requested summaries.
}
\description{
The \code{summary} method provides summary statistics regarding
the performance of \code{cvCovEst} and can be used for diagnostic plotting.
}
