% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_cvCovEst}
\alias{theme_cvCovEst}
\title{cvCovEst Plot Theme}
\usage{
theme_cvCovEst(plot_type)
}
\arguments{
\item{plot_type}{A character vector}

\item{plot_title}{}

\item{...}{Additional arguments to be passed to \code{cvCovEst} plotting
functions.}
}
\value{
A \code{ggplot} theme.
}
\description{
\code{theme_cvCovEst} defines the overall theme of the
\code{cvCovEst} plotting functions and makes changes depending on which plot
function is being called.
}
\keyword{internal}
