% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{thresholdingEst}
\alias{thresholdingEst}
\title{Simple Thresholding Estimator of Covariance Matrix}
\usage{
thresholdingEst(data, gamma)
}
\arguments{
\item{data}{A numeric \code{data.frame} or matrix.}

\item{gamma}{A \code{numeric} larger than or equal to 0 defining the hard
threshold applied to each element of the \code{data}'s sample covariance
matrix.}
}
\value{
A \code{matrix} corresponding to the estimate of the covariance
matrix.
}
\description{
\code{thresholdingEst} computes the thresholding estimate of the
covariance matrix for a given value of \code{gamma}. The simple threshold
estimator of convariance matrix applies a hard thresholding operator to
each element of the sample covariance matrix.
}
